Separated firebaseconfig to its own folder for better useage
